pyproject.toml
[tool.poetry]
name = "repl_python3_realbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"dislash.py" = "^1.4.9"
aioconsole = "^0.3.3"
Flask = "^2.0.2"
art = "^5.3"
discord = "^1.7.3"
nextcord = "^2.0.0-alpha.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
